+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support for application definition parameters update in dataflow application
 - Support for Cross Region Replication
 - Support for Secrets Management Service's `oci_vault_secret` and `oci_vault_secret_version` datasources
+- Support for Retention Rules that control object immutability
 
 ## 3.68.0 (March 25, 2020)